Big optimization of SMBX64 formats readers Reading performance increased into 2.0~3.0 times!

+++ lvl_filedata.h
@@ -810,7 +810,7 @@ struct LevelData
     //! Line data where error was occouped
     PGESTRING ERROR_linedata;
     //! Number of line where error was occouped
-    int       ERROR_linenum;
+    long      ERROR_linenum;
     /*!
      * \brief File format
      */
